DA SYKA (in.cpp bug fix + GRB rules changes)

+++ LP_LIB/LP_LIB.cpp
@@ -37,7 +37,7 @@ extern "C" {
 		if (buffer == NULL)
 			return NULL;
 
-		int bin = 0, i = 1;
+		unsigned long bin = 0, i = 1;
 		while (number) {
 			bin += (number % 2) * i;
 			i *= 10;
